separate verify tests per class
