+++ test/verify/test_arg_fusion.cpp
@@ -121,81 +121,3 @@ template struct test_arg_fusion<migraphx::op::argmax, migraphx::shape::half_type
 // no fusion (just arg op) tests
 template struct test_arg_fusion<migraphx::op::argmin, migraphx::shape::float_type, 1, 5>;
 template struct test_arg_fusion<migraphx::op::argmax, migraphx::shape::float_type, 1, 5>;
-
-// Additional tests with variable shapes (don't fit template pattern)
-
-// Test argmin on axis 0 with different shape {4, 3, 2}
-struct test_arg_fusion_argmin_shape_4_3_2 : verify_program<test_arg_fusion_argmin_shape_4_3_2>
-{
-    migraphx::program create_program() const
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {4, 3, 2}};
-        auto x    = mm->add_parameter("x", s);
-        auto relu = mm->add_instruction(migraphx::make_op("relu"), x);
-        mm->add_instruction(migraphx::make_op("argmin", {{"axis", 0}}), relu);
-        return p;
-    }
-
-    std::string section() const { return "reduce"; }
-};
-
-// Test argmax on last axis with larger reduction {8, 5, 256}
-struct test_arg_fusion_argmax_shape_8_5_256 : verify_program<test_arg_fusion_argmax_shape_8_5_256>
-{
-    migraphx::program create_program() const
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {8, 5, 256}};
-        auto x   = mm->add_parameter("x", s);
-        auto neg = mm->add_instruction(migraphx::make_op("neg"), x);
-        mm->add_instruction(migraphx::make_op("argmax", {{"axis", -1}}), neg);
-        return p;
-    }
-
-    std::string section() const { return "reduce"; }
-};
-
-// Test argmin with broadcast input fusion
-struct test_arg_fusion_argmin_broadcast : verify_program<test_arg_fusion_argmin_broadcast>
-{
-    migraphx::program create_program() const
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 4}};
-        migraphx::shape bs{migraphx::shape::float_type, {1, 3, 4}};
-        auto x = mm->add_parameter("x", s);
-        auto y = mm->add_parameter("y", bs);
-        auto yb =
-            mm->add_instruction(migraphx::make_op("multibroadcast", {{"out_lens", s.lens()}}), y);
-        auto add = mm->add_instruction(migraphx::make_op("add"), x, yb);
-        mm->add_instruction(migraphx::make_op("argmin", {{"axis", 1}}), add);
-        return p;
-    }
-
-    std::string section() const { return "reduce"; }
-};
-
-// Test argmax with broadcast input fusion
-struct test_arg_fusion_argmax_broadcast : verify_program<test_arg_fusion_argmax_broadcast>
-{
-    migraphx::program create_program() const
-    {
-        migraphx::program p;
-        auto* mm = p.get_main_module();
-        migraphx::shape s{migraphx::shape::float_type, {2, 3, 4}};
-        migraphx::shape bs{migraphx::shape::float_type, {2, 1, 4}};
-        auto x = mm->add_parameter("x", s);
-        auto y = mm->add_parameter("y", bs);
-        auto yb =
-            mm->add_instruction(migraphx::make_op("multibroadcast", {{"out_lens", s.lens()}}), y);
-        auto mul = mm->add_instruction(migraphx::make_op("mul"), x, yb);
-        mm->add_instruction(migraphx::make_op("argmax", {{"axis", 2}}), mul);
-        return p;
-    }
-
-    std::string section() const { return "reduce"; }
-};
